requirements.txt
requests==2.31.0
SpeechRecognition==3.10.0
pyttsx3==2.90
python-telegram-bot==21.5
flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.0.1
